+++ system/cloud/gcp/storage/usage/hello.go
@@ -13,17 +13,15 @@ type PubsubProxyEvent struct {
 	Attributes *Attributes
 }
 
-
 type Attributes struct {
-	NotificationConfig string `json:"notificationConfig"`
-	ObjectGeneration string `json:"objectGeneration"`
-	ObjectId string  `json:"objectId"`
-	BucketId string `json:"bucketId"`
-	EventTime  *time.Time `json:"eventTime"`
-	EventType string `json:"eventTime"`
-	OverwroteGeneration  string `json:"overwroteGeneration"`
+	NotificationConfig  string     `json:"notificationConfig"`
+	ObjectGeneration    string     `json:"objectGeneration"`
+	ObjectId            string     `json:"objectId"`
+	BucketId            string     `json:"bucketId"`
+	EventTime           *time.Time `json:"eventTime"`
+	EventType           string     `json:"eventTime"`
+	OverwroteGeneration string     `json:"overwroteGeneration"`
 }
-
 
 //Hello prints meta and events details
 func Hello(ctx context.Context, event *PubsubProxyEvent) error {