+++ testing/endpoint/http/init.go
@@ -10,5 +10,3 @@ func init() {
 	})
 	initProviders()
 }
-
-